move location and numpybatch
