+++ ocf_datapipes/batch/__init__.py
@@ -11,3 +11,4 @@ from .merge_numpy_examples_to_batch import (
 )
 from .merge_numpy_modalities import MergeNumpyModalitiesIterDataPipe as MergeNumpyModalities
 from .merge_numpy_modalities import MergeNWPNumpyModalitiesIterDataPipe as MergeNWPNumpyModalities
+from .batches import BatchKey, NumpyBatch, NWPBatchKey, NWPNumpyBatch, XarrayBatch
