+++ ocf_datapipes/batch/fake/sun.py
@@ -3,7 +3,7 @@ import numpy as np
 
 from ocf_datapipes.batch.fake.utils import get_n_time_steps_from_config
 from ocf_datapipes.config.model import Configuration
-from ocf_datapipes.utils.consts import BatchKey, NWPBatchKey
+from ocf_datapipes.batch import BatchKey, NWPBatchKey
 
 
 def make_fake_sun_data(configuration: Configuration, batch_size: int = 8):
