+++ ocf_datapipes/batch/merge_numpy_modalities.py
@@ -2,7 +2,7 @@
 from torch.utils.data import IterDataPipe, functional_datapipe
 
 from ocf_datapipes.utils import Zipper
-from ocf_datapipes.utils.consts import BatchKey, NumpyBatch
+from ocf_datapipes.batch import BatchKey, NumpyBatch
 
 
 @functional_datapipe("merge_numpy_modalities")
