+++ ocf_datapipes/convert/numpy_batch/gsp.py
@@ -3,7 +3,7 @@ import logging
 
 from torch.utils.data import IterDataPipe, functional_datapipe
 
-from ocf_datapipes.utils.consts import BatchKey, NumpyBatch
+from ocf_datapipes.batch import BatchKey, NumpyBatch
 from ocf_datapipes.utils.utils import datetime64_to_float
 
 logger = logging.getLogger(__name__)
