+++ ocf_datapipes/convert/numpy_batch/satellite.py
@@ -1,7 +1,7 @@
 """Convert Satellite to NumpyBatch"""
 from torch.utils.data import IterDataPipe, functional_datapipe
 
-from ocf_datapipes.utils.consts import BatchKey, NumpyBatch
+from ocf_datapipes.batch import BatchKey, NumpyBatch
 from ocf_datapipes.utils.utils import datetime64_to_float
