+++ ocf_datapipes/utils/__init__.py
@@ -3,3 +3,4 @@ from .datapipes import LengthSetterIterDataPipe as LengthSetter
 from .datapipes import RepeaterIterDataPipe as Repeater
 from .datapipes import UnZipperIterDataPipe as UnZipper
 from .datapipes import ZipperIterDataPipe as Zipper
+from .location import Location
