+++ tests/training/example/test_pv_nwp.py
@@ -1,5 +1,5 @@
 from ocf_datapipes.training.example.pv_nwp import pv_nwp_datapipe
-from ocf_datapipes.utils.consts import BatchKey
+from ocf_datapipes.batch import BatchKey
 
 
 def test_pv_nwp_datapipe(configuration_with_pv_netcdf_and_nwp):
