+++ tests/training/example/test_simple_pv.py
@@ -1,5 +1,5 @@
 from ocf_datapipes.training.example.simple_pv import simple_pv_datapipe
-from ocf_datapipes.utils.consts import BatchKey
+from ocf_datapipes.batch import BatchKey
 
 
 def test_simple_pv_datapipe(configuration_with_pv_netcdf):
